refactor: split tabs into components
